타입 변경 및 Repository에서 optional을 내려주는것으로 변경

+++ src/main/java/com/devagit/springbootstudy/domain/Heart.java
@@ -16,7 +16,7 @@ public class Heart {
     @GeneratedValue
     @Id
     private long id;
-    @Column(columnDefinition = "long")
+    @Column(columnDefinition = "INT")
     private long postId;
     @Column(columnDefinition = "VARCHAR(32)", nullable = false)
     private String userId;
